Fix web worker usage in account report
