+++ app/features/TransactionSlice.ts
@@ -48,7 +48,7 @@ import { GetTransactionsOutput } from '~/preload/preloadTypes';
 import { findEntries } from '~/database/DecryptedAmountsDao';
 import { getActiveBooleanFilters } from '~/utils/accountHelpers';
 import errorMessages from '~/constants/errorMessages.json';
-import { isSuccessfulEncryptedTransaction } from '~/utils/decryptHelpers';
+import isSuccessfulEncryptedTransaction from '~/utils/decryptHelpers';
 
 export const transactionLogPageSize = 100;
