+++ app/utils/types.ts
@@ -555,6 +555,10 @@ export interface TransferTransactionWithNames extends TransferTransaction {
     toName?: string;
 }
 
+export interface DecryptedTransferTransaction extends TransferTransaction {
+    decryptedAmount: string;
+}
+
 export type EncryptedAmount = Hex;
 
 export interface TypedCredentialDeploymentInformation {
